Optimize &CONSTSUB
Calling a constant subroutine with & breaks the optimization; this
commit attempts to perform it anyway if the constant is known at
compile time.

As a side effect, the module now has to meddle with defined() and
exists(), which special-case &sub; codewise, this means that

exists &FOO

and

defined &FOO

can now both be optimized out early.